add statements.c/h
Now we create at first a high level intermediate representation
using the struct statement_t. This struct will be translated into
ir_node_t which is the low level intermediate representation.

This is a first step to implement more sophisticated implementations
afterwards.

Thanks to Vasilis Vasiliadis for the initial structs.