final_code.c
merge register allocator

Well, I forgot to initialize the physical register pool, that's why
it wasn't working all that time. Also fixed an out of bound pool access
for some registers which was responsible for some weird behavior of
the allocator.

Tested only for integer instructions, no spilling yet.

(Note)
The register allocation is the last pass of the code, for this reason
the code lives in final_code.c and not in a seperate file. Maybe when
it will get more sophisticated I will move it alone. But some code
reorganization must be done first.