fix variable analysis bug
Subprograms (even totaly unrelated between each other) when use
variables from previous scopes do not restore their undefined
state. This was fatal for some optimizations emiting incorrect code.

Also put the first assignment of an uninitialized variable in main
into the .data segmet. It was implemented but disabled due to this
bug. Now we can turn this optimization on.

This was another core bug, hopefully it was much easier to fix it,
although this is a very "play safe" implementation.