implement block/statement analysis
For every statement we keep a list of variable lists for reading
and a list of variables for writing

For every block we keep all the above information for every statement
in this block. We have lots of duplicate information for now, more
cleaning will be done later, depending which form of statistics fits
better to adta analysis.

We have lots of debugging output at the moment for testing